Bump default board size to 13x13
Most computers are fast enough for this, and 9x9 highlights our lack
of symmetry and is generally too small for many aspects to show up.

We might be currently around GNUGo level on 13x13 2000sims, though
that is very rough estimate - we are certainly in the same ballpark
and Michi seems to tend to win a lot w/ 4000sims (then even on 15x15).