Add large-scale pattern support
This is a big performance hit in the tree thread, which is a little
inconvenient - we needed to bump up the expand threshold a lot and
lower the total #simulations to compensate, but this should still
be totally worth it.

We simply use Pachi data files without any modification.

Fitting this within the 500-line goal will be a challenge, though!